+++ .vercel/output/_functions/chunks/_virtual_astro_get-image_T83WRX-z.mjs
@@ -1,0 +1,12 @@
+import { g as getImage$1 } from './deterministic-string_D3tbQ0ks.mjs';
+import './entrypoint_OVoc7678.mjs';
+
+const imageConfig = {"endpoint":{"route":"/_image"},"service":{"entrypoint":"astro/assets/services/sharp","config":{}},"dangerouslyProcessSVG":false,"domains":[],"remotePatterns":[],"responsiveStyles":false};
+								Object.defineProperty(imageConfig, 'assetQueryParams', {
+									value: undefined,
+									enumerable: false,
+									configurable: true,
+								});
+								const getImage = async (options) => await getImage$1(options, imageConfig);
+
+export { getImage, imageConfig };