+++ .vercel/output/_functions/chunks/creativity-constraints_CLgyX-et.mjs
@@ -1,0 +1,8 @@
+async function getMod() {
+						return import('./creativity-constraints_RmFLheGy.mjs');
+					}
+					const collectedLinks = [];
+					const collectedStyles = [];
+					const defaultMod = { __astroPropagation: true, getMod, collectedLinks, collectedStyles, collectedScripts: [] };
+
+export { defaultMod as default };