+++ .vercel/output/_functions/entry.mjs
@@ -1,0 +1,1 @@
+export { P as default } from './chunks/entrypoint_OVoc7678.mjs';